requirements.txt
fastapi==0.95.2
# FastAPI 0.95 仅兼容 Pydantic v1；勿与父目录 .venv 中已升级的 Pydantic 2 混用
pydantic>=1.10.13,<2
uvicorn[standard]==0.22.0
motor==3.3.1
pymongo==4.6.1
python-jose[cryptography]==3.3.0
bcrypt==4.1.3
python-multipart==0.0.6
pytest==7.4.0
pytest-asyncio==0.21.1
httpx==0.24.1
